+++ src/main/java/com/myEducation/inside/dao/UserDao.java
@@ -8,7 +8,7 @@ import com.myEducation.inside.model.User;
 import com.myEducation.inside.utils.baseDao.BaseDao;
 import com.myEducation.inside.utils.page.Page;
 
-public interface UserDao extends BaseDao<User, Long>{
+public interface UserDao extends BaseDao<User, Integer>{
 	
 	// 分页查询显示
 	public List<User> selectByPage(Page<User> page);
